backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy>=2.0.0
asyncpg
pydantic-settings
pydantic>=2.0.0
alembic
celery
redis[asyncio]
boto3
httpx
tiktoken
openai
anthropic
pyyaml
psycopg[binary]
pytest
pytest-asyncio
# platform layer
python-jose[cryptography]
passlib[bcrypt]
stripe>=8.0.0
pydantic[email]
